Use refactored Registers types
This change switches from Register/Registers from helix-core to the
new registers types introduced in the parent commits. This is an
unfortunately noisy change: we need to update virtually all callsites
that access the registers.

We need to move Registers out of the Editor type to make this work:
'RegisterValues''s lifetime is based on the Register and the Editor
lifetime. We move Registers up to the Context types to resolve this.