registers: Preview the latest value for simple registers
This fixes a discrepancy between simple registers which are used for
yanking multiple values (for example yanking via `"ay`) and simple
registers that store a history of values (for example `"a*`).

Previously, the preview shown in `select_register`'s infobox would show
the oldest value in history. It's intuitive and useful to see the most
recent value pushed to the history though.

We cannot simply switch the preview line from `values.first()` to
`values.last()`: that would fix the preview for registers used for
history but break the preview for registers used to yank multiple
values. We could push to the beginning of the values with
`SimpleRegister::push` but this is wasteful from a performance
perspective. Instead we can have `SimpleRegister::read` return an
iterator that returns elements in the reverse order and reverse the
values in `SimpleRegister::write`. This effectively means that `push`
adds elements to the beginning of the register's values. For the sake
of the preview, we can switch to `values.last()` and that is then
correct for both usage-styles of `SimpleRegister`. This also needs a
change to call-sites that read the latest history value to switch from
`last` to `first`.